업로드 파이프라인 뼈대 및 admin 스크립트
- /upload 요청 바디를 파일로 저장하고 썸네일 생성·DB 등록·세그먼트 생성까지 수행하는 기본 처리 추가(단순 multipart 전제)
- ffmpeg 기반 처리 후 JSON으로 video_id 반환
- tools/admin.sh에 upload 헬퍼 추가

+++ src/server/upload.c
@@ -7,6 +7,7 @@
 #include <string.h>
 #include <sys/stat.h>
 #include <unistd.h>
+#include <time.h>
 
 static int ensure_dir(const char *path) {
     if (access(path, F_OK) == 0) {
@@ -37,7 +38,7 @@ int handle_upload_request(int fd, const char *headers, const char *body, size_t 
 
     int video_id = 0;
     char base_name[64];
-    snprintf(base_name, sizeof(base_name), "video_%ld", time(NULL));
+    snprintf(base_name, sizeof(base_name), "video_%ld", (long)time(NULL));
 
     char video_path[256];
     snprintf(video_path, sizeof(video_path), "data/videos/%s.mp4", base_name);
